Move Measurement and TelemetryPacket structs from proc to tlm lib to avoid future cyclic dependency issues
